split tiles into their own json files
