+++ components/MovieTile.js
@@ -1,0 +1,19 @@
+import React from "react";
+import { View, Image, Text } from "react-native";
+
+const MovieTile = ({ item }) => {
+    const { image, fullTitle, imdbRating, imDbRatingCount } = item;
+    return (
+      <View style={styles.movieScreenInfo}>
+        <View style={styles.movieScreenImageView}>
+          <Image source={{ uri: image }} style={styles.movieScreenImage} />
+        </View>
+        <Text style={styles.fullTitle}>{fullTitle}</Text>
+        <Text style={styles.imdbRating}>
+          IMDB Rating: {imdbRating} ({imDbRatingCount} reviews)
+        </Text>
+      </View>
+    );
+  };
+
+  export default MovieTile
