+++ screens/ComingSoonScreen.js
@@ -2,7 +2,7 @@ import React from "react";
 import { SafeAreaView } from "react-native";
 
 import styles from "../styles";
-import { ComingSoonTile } from "../components/Tile";
+import ComingSoonTile from "../components/ComingSoonTile";
 
 const ComingSoonScreen = ({ route }) => {
   return (
