+++ screens/MovieScreen.js
@@ -2,12 +2,12 @@ import React from "react";
 import { SafeAreaView } from "react-native";
 
 import styles from "../styles";
-import { RatingTile } from "../components/Tile";
+import MovieTile from "../components/MovieTile";
 
 const MovieScreen = ({ route }) => {
   return (
     <SafeAreaView style={styles.movieScreenRoot}>
-      <RatingTile item={route.params} />
+      <MovieTile item={route.params} />
     </SafeAreaView>
   );
 };
